Fix wrong assertion and refactor tests
